fix(alloc): the per-class cap's u16 was a hidden 4 GiB ceiling — the first soak found it
The runaway guard had already been raised once ("three orders of
magnitude", the doc records) but its own u16 pinned it at 65,535
spans x 64 KiB = 4 GiB per size class. The first hour-long soak
reached it in ten minutes: a 3-byte-value mixed storm filled the 16 B
class and the process aborted with "memory allocation of 3 bytes
failed" on a box with 48 GiB free. Counter and cap are now u32 with
the guard at 1 TiB per class - memory governance belongs to maxmemory
and the tier budget, never to an invisible allocator constant. The
exhaustion path stays testable via with_class_cap.

kevy-alloc 41 tests green; workspace green.

+++ crates/kevy-alloc/src/tests.rs
@@ -147,7 +147,7 @@ fn m6_an_exhausted_class_refuses_instead_of_handing_back_a_wild_pointer() {
     // The default cap is a runaway guard set past any real workload, so
     // a tighter one is used here to reach the refusal at all — see
     // `PER_CLASS_CAP` for why the inherited value was wrong.
-    const CAP: u16 = 3;
+    const CAP: u32 = 3;
     let mut heap = Heap::with_class_cap(0, CAP);
     // The class is chosen large so the number of allocations stays
     // small: 8192-byte slots give 8 per span.
